change dtype of Lattice mapping scale matrix to np.int
Former-commit-id: b82fe98337d483dfd2e0930857b88a30d2b54d8a [formerly 630308175fa44b5feae2ff71410ddbdcc5c39927]
Former-commit-id: a9363122826e11ac34778763745dbfd85f1e8557

+++ pymatgen/core/lattice.py
@@ -510,7 +510,7 @@ class Lattice(MSONable):
                                   np.logical_and(alphab,
                                                  betab[i][None, :]))
             for j, k in np.argwhere(inds):
-                scale_m = np.array((f_a[i], f_b[j], f_c[k]))
+                scale_m = np.array((f_a[i], f_b[j], f_c[k]), dtype=np.int)
                 if abs(np.linalg.det(scale_m)) < 1e-8:
                     continue
